requirements.txt
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
kaleido>=0.2.1
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
pydantic>=2.5.0
psycopg2-binary>=2.9.9
fpdf2>=2.7.0